TranscriptionExperiments/requirements.txt
SpeechRecognition
pyaudio
pocketsphinx
azure-cognitiveservices-speech